Changed storage time to storage quantity per web hook, and added more rigorous testing as this is a bit more complicated. Also separated the log tests.
