+++ tests/CaptainHookTest.php
@@ -48,7 +48,7 @@ class CaptainHookTest extends Orchestra\Testbench\TestCase
         });
         $app['config']->set('captain_hook.listeners', ['eloquent.*']);
         $app['config']->set('captain_hook.log.active', true);
-        $app['config']->set('captain_hook.log.storage_time', 24);
+        $app['config']->set('captain_hook.log.storage_quantity', 50);
         $app['config']->set('database.default', 'testing');
         $app['config']->set('queue.driver', 'sync');
 
@@ -239,157 +239,6 @@ class CaptainHookTest extends Orchestra\Testbench\TestCase
         $obj = new TestModel();
         $obj->name = 'Test';
         $obj->save();
-    }
-
-    public function testItRunsSynchronouslyInJobWhileLogging()
-    {
-        $this->app['config']->set('queue.driver', 'notSync');
-
-        $provider = $this->app->getProvider('Mpociot\\CaptainHook\\CaptainHookServiceProvider');
-
-        $webhook = new Webhook();
-        $webhook->url = 'http://test.foo/saved';
-        $webhook->event = 'eloquent.saved: TestModel';
-        $webhook->save();
-
-        $webhook = new Webhook();
-        $webhook->url = 'http://test.bar/saved';
-        $webhook->event = 'eloquent.saved: TestModel';
-        $webhook->save();
-
-        $webhook = new Webhook();
-        $webhook->url = 'http://test.foo/deleted';
-        $webhook->event = 'eloquent.deleted: TestModel';
-        $webhook->save();
-
-        $client = m::mock('GuzzleHttp\\Client');
-
-        $client->shouldReceive('getConfig')
-            ->with('handler')
-            ->andReturn(\GuzzleHttp\Middleware::httpErrors());
-
-        $client->shouldReceive('post')
-            ->twice();
-
-        $client->shouldReceive('post')
-            ->with('http://test.foo/saved', m::any());
-
-        $client->shouldReceive('post')
-            ->with('http://test.bar/saved', m::any());
-
-        $this->app->instance(GuzzleHttp\Client::class, $client);
-
-        $provider->setClient($client);
-        $provider->setWebhooks(Webhook::all());
-
-        $obj = new TestModel();
-        $obj->name = 'Test';
-        $obj->save();
-    }
-
-    public function testItLogsTriggeredWebhooks()
-    {
-        $this->app['config']->set('queue.driver', 'notSync');
-
-        $provider = $this->app->getProvider('Mpociot\\CaptainHook\\CaptainHookServiceProvider');
-
-        $webhook = new Webhook();
-        $webhook->url = 'http://test.foo/saved';
-        $webhook->event = 'eloquent.saved: TestModel';
-        $webhook->save();
-
-        $webhook = new Webhook();
-        $webhook->url = 'http://test.bar/saved';
-        $webhook->event = 'eloquent.saved: TestModel';
-        $webhook->save();
-
-        $webhook = new Webhook();
-        $webhook->url = 'http://test.foo/deleted';
-        $webhook->event = 'eloquent.deleted: TestModel';
-        $webhook->save();
-
-        $handler = new \GuzzleHttp\Handler\MockHandler([
-            new \GuzzleHttp\Psr7\Response(200, [
-                'Content-Type' => 'application/json',
-            ], '{"data":"First data"}'),
-            new \GuzzleHttp\Psr7\Response(200, [
-                'Content-Type' => 'application/json',
-            ], '{"data":"Second data"}'),
-        ]);
-
-        $handler = \GuzzleHttp\HandlerStack::create($handler);
-
-        $client = new \GuzzleHttp\Client(['handler' => $handler]);
-
-        $this->app->instance(GuzzleHttp\Client::class, $client);
-
-        $provider->setClient($client);
-
-        $test = new TestModel();
-        $test->name = 'Test';
-        $test->save();
-
-        $this->seeInDatabase('captain_hook_logs', [
-            'webhook_id' => 1,
-            'url' => 'http://test.foo/saved',
-            'response' => '{"data":"First data"}',
-            'response_format' => 'application/json',
-        ]);
-
-        $this->seeInDatabase('captain_hook_logs', [
-            'webhook_id' => 2,
-            'url' => 'http://test.bar/saved',
-            'response' => '{"data":"Second data"}',
-            'response_format' => 'application/json',
-        ]);
-    }
-
-    public function testItRemovesOldWebhooks()
-    {
-        $this->app['config']->set('queue.driver', 'notSync');
-        $this->app['config']->set('captain_hook.listeners', ['eloquent.saved: TestModel']);
-
-        $webhook = new Webhook();
-        $webhook->url = 'http://test.foo/saved';
-        $webhook->event = 'eloquent.saved: TestModel';
-        $webhook->save();
-
-        $log = new \Mpociot\CaptainHook\WebhookLog([
-            'webhook_id' => null,
-            'url' => 'anything',
-            'payload_format' => null,
-            'payload' => '',
-            'status' => 200,
-            'response' => '',
-            'response_format' => null,
-        ]);
-        $log->created_at = $log->updated_at = date('Y-m-d H:i:s', strtotime('-23 hours'));
-        $log->timestamps = false;
-
-        $log->save();
-
-        $this->seeInDatabase('captain_hook_logs', [
-            'webhook_id' => null,
-            'url' => 'anything',
-            'payload_format' => null,
-            'payload' => '',
-            'status' => 200,
-            'response' => '',
-            'response_format' => null,
-        ]);
-
-        $log->updated_at = date('Y-m-d H:i:s', strtotime('-25 hours'));
-        $log->save();
-
-        $this->dontSeeInDatabase('captain_hook_logs', [
-            'webhook_id' => null,
-            'url' => 'anything',
-            'payload_format' => null,
-            'payload' => '',
-            'status' => 200,
-            'response' => '',
-            'response_format' => null,
-        ]);
     }
 }
